Still untested changes.. Rearranged some more.  We are now calling the 2 main functions in controller.c.  Note however that the rate of polling of controller_handle() is not as fast as we had decided earlier.  It is presently called every 10mS instead of the 3 or 4mS we talked about.
Enabled the MIDI_COMM_ROUTINE() to allow midi data to be sent out.  (Needs to be tested)

Further, the expression pedal data is now being fed into the expression pedal lookup table to convert it to a linear value.  Note that the table is set up completely linear right now until I find my data on this..

Rearranged ram to utilize the access bank.

Commented out all of the test code at the beginning of main().  

This code is ready to be actually tested.  I believe everything is ready to go.

git-svn-id: svn://localhost/e-minor-controller/trunk@40 81a870ca-24fd-4f65-b8a6-cbdfd7b1111e