Jim seems to think that an int is always 32 bits...  Changed the types.h to indicate u32 as a unsigned long instead of int.  Thats why the device was going bonkers and appearing to be plotting the destruction of the planet.
Changed the 7seg display to update in the isr to ensure that it is updated often enough.

Added the data from eeprom.bin as test data.

The function that displays the ascii characters on the display is backwards for the controller code.  Instead of Soc, its coS.

Something else is wierd.. in certain modes the unit appears to bomb and just plain go nuts.  I'm not sure if this is a noise issue or a code issue.  I'll probably add some caps to the buttons and see if that helps.

git-svn-id: svn://localhost/e-minor-controller/trunk@47 81a870ca-24fd-4f65-b8a6-cbdfd7b1111e