Something wierd was happening causing the unit to write to program memory which was screwing everything up.  I've disabled the write code for now.
Fixed the order of displayed characters on the display.

Fixed the unit thinking the buttons were always held down.  ButtonState has been changed to an unsigned char instead of an unsigned long.  Now we're acutally representing the button data to the controller code properly.  

This appears to be a bit more stable... though something is still funky and I'm not sure what it is..

git-svn-id: svn://localhost/e-minor-controller/trunk@48 81a870ca-24fd-4f65-b8a6-cbdfd7b1111e