Removed all of the defined const variables in code as they doesn't work the way we've got the microcontroller set up.  The compiler is kinda stupid.  It could have worked if they were also defined as rom, but alas.. meh.. I've converted them all to #defines.
Also the ram data initialization doesn't work either for the same reason the const didn't.  I've put all ram initializations into controller_init().

Added a MODE_UNDEFINED to handle the startup case when we don't know the mode.  This reduces the number of paths to get to a given mode.

Right now the problem is that leds_show_4alphas is being called from 2 different locations which are attempting to put different data into the display.  The result is that the display is switching back and fourth between the two rapidly.

Disabled the 10ms controller timing code for testing..

git-svn-id: svn://localhost/e-minor-controller/trunk@49 81a870ca-24fd-4f65-b8a6-cbdfd7b1111e