Fixed more bitchouts.  Fixed timing stuff..
Fixed the display of SEQ-.

I wired it wrong so I had to reverse the positions of INC and NEXT buttons in firmware.

Enabled the watchdog, and brownout reset..

Added a mode led to show practice/concert.  Concert has the last decimal point on the right lit.

Re-enabled the FLASH write code which causes the unit to bomb upon trying it.

git-svn-id: svn://localhost/e-minor-controller/trunk@51 81a870ca-24fd-4f65-b8a6-cbdfd7b1111e